% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-methods.R
\name{delete_project}
\alias{delete_project}
\title{Delete a project}
\usage{
delete_project(
  project_name,
  delete_experiments = TRUE,
  workspace_name = NULL,
  api_key = NULL
)
}
\arguments{
\item{project_name}{Project name.}

\item{delete_experiments}{If \code{TRUE}, delete all the experiments in the project.}

\item{workspace_name}{Workspace name (can also be specified using the \code{COMET_WORKSPACE}
parameter as an environment variable or in a comet config file).}

\item{api_key}{Comet API key (can also be specified using the \code{COMET_API_KEY}
parameter as an environment variable or in a comet config file).}
}
\description{
Delete a project
}
\examples{
\dontrun{
library(cometr)
# Assuming you have COMET_API_KEY, COMET_WORKSPACE variables defined
delete_project(project_name = "project1")
}

}
