% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logging.R
\name{disable_logging}
\alias{disable_logging}
\title{Disable cometr logging}
\usage{
disable_logging()
}
\description{
Generally, if the \code{COMET_LOGGING_FILE} and \code{COMET_LOGGING_FILE_LEVEL} parameters
are found, then \code{cometr} will log internal information. You can disable logging
for a particular R session by calling \code{disable_logging()}.
}
