% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{experiment}
\alias{experiment}
\title{Create a new experiment}
\usage{
experiment(
  project_name = NULL,
  workspace = NULL,
  api_key = NULL,
  log_errors = FALSE
)
}
\description{
Create a new experiment
}
