% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_general.R
\name{experiments}
\alias{experiments}
\title{Get a project's experiments}
\usage{
experiments(
  project_id = NULL,
  project_name = NULL,
  workspace_name = NULL,
  archived = FALSE,
  api_key = NULL
)
}
\arguments{
\item{project_id}{Project ID.}

\item{project_name}{Project name.}

\item{workspace_name}{Workspace name.}

\item{archived}{Whether to retrieve archives experiments.}

\item{api_key}{Comet API key.}
}
\description{
Either \code{project_id} should be provided, or both \code{project_name} and \code{workspace_name}
should be provided.
}
