% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-methods.R
\name{get_workspaces}
\alias{get_workspaces}
\title{Get a user's workspaces}
\usage{
get_workspaces(api_key = NULL)
}
\arguments{
\item{api_key}{Comet API key (can also be specified using the \code{COMET_API_KEY}
parameter as an environment variable or in a comet config file).}
}
\description{
Get a user's workspaces
}
\examples{
\dontrun{
library(cometr)
# Assuming you have COMET_API_KEY variable defined
get_workspaces()
}

}
