% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logged-artifact.R
\name{LoggedArtifact}
\alias{LoggedArtifact}
\title{A Logged Comet Artifact object}
\description{
Comet Artifacts allow keeping track of assets beyond any particular experiment.
The \code{LoggedArtifact} is a Comet \code{\link{Artifact}} that already logged to the Comet servers
and can be used to access the artifact version assets and download them locally.
}
\examples{
\dontrun{
library(cometr)
# Assuming you have COMET_API_KEY, COMET_WORKSPACE, COMET_PROJECT_NAME variables define
exp <- create_experiment()

# Get a Comet Artifact
artifact <- exp$get_artifact(artifact_name = "workspace/artifact-name:versionOrAlias")

exp$stop()
}

}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-LoggedArtifact-new}{\code{LoggedArtifact$new()}}
\item \href{#method-LoggedArtifact-get_artifact_name}{\code{LoggedArtifact$get_artifact_name()}}
\item \href{#method-LoggedArtifact-get_artifact_type}{\code{LoggedArtifact$get_artifact_type()}}
\item \href{#method-LoggedArtifact-get_artifact_version}{\code{LoggedArtifact$get_artifact_version()}}
\item \href{#method-LoggedArtifact-get_aliases}{\code{LoggedArtifact$get_aliases()}}
\item \href{#method-LoggedArtifact-get_metadata}{\code{LoggedArtifact$get_metadata()}}
\item \href{#method-LoggedArtifact-get_version_tags}{\code{LoggedArtifact$get_version_tags()}}
\item \href{#method-LoggedArtifact-assets}{\code{LoggedArtifact$assets()}}
\item \href{#method-LoggedArtifact-remote_assets}{\code{LoggedArtifact$remote_assets()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoggedArtifact-new"></a>}}
\if{latex}{\out{\hypertarget{method-LoggedArtifact-new}{}}}
\subsection{Method \code{new()}}{
Creates new \code{LoggedArtifact} object with provided parameters. Do not use
this method directly. Use \code{Experiment$get_artifact()}
to retrieve \code{LoggedArtifact}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoggedArtifact$new(
  artifact_name,
  artifact_type,
  artifact_id,
  artifact_version_id,
  workspace,
  experiment_key,
  artifact_version,
  aliases,
  artifact_tags,
  version_tags,
  size,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{artifact_name}}{(Required) Artifact name.}

\item{\code{artifact_type}}{(Required) The artifact type.}

\item{\code{artifact_id}}{(Required) The ID of artifact.}

\item{\code{artifact_version_id}}{(Required) The ID of Artifact Version.}

\item{\code{workspace}}{(Required) The workspace where artifact saved}

\item{\code{experiment_key}}{(Required) The ID of associated experiment}

\item{\code{artifact_version}}{(Required) The latest artifact version.}

\item{\code{aliases}}{(Required) List of Artifact Version aliases.}

\item{\code{artifact_tags}}{(Required) The list of artifact tags.}

\item{\code{version_tags}}{(Required) List of Artifact Version tags.}

\item{\code{size}}{(Required) The total size of logged artifact version.
It is the sum of all the artifact version assets.}

\item{\code{metadata}}{The meta-data of Artifact Version.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoggedArtifact-get_artifact_name"></a>}}
\if{latex}{\out{\hypertarget{method-LoggedArtifact-get_artifact_name}{}}}
\subsection{Method \code{get_artifact_name()}}{
Get the name of the artifact.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoggedArtifact$get_artifact_name()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoggedArtifact-get_artifact_type"></a>}}
\if{latex}{\out{\hypertarget{method-LoggedArtifact-get_artifact_type}{}}}
\subsection{Method \code{get_artifact_type()}}{
Get the type of the artifact.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoggedArtifact$get_artifact_type()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoggedArtifact-get_artifact_version"></a>}}
\if{latex}{\out{\hypertarget{method-LoggedArtifact-get_artifact_version}{}}}
\subsection{Method \code{get_artifact_version()}}{
Get the version of the artifact.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoggedArtifact$get_artifact_version()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoggedArtifact-get_aliases"></a>}}
\if{latex}{\out{\hypertarget{method-LoggedArtifact-get_aliases}{}}}
\subsection{Method \code{get_aliases()}}{
Get the version of the artifact.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoggedArtifact$get_aliases()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoggedArtifact-get_metadata"></a>}}
\if{latex}{\out{\hypertarget{method-LoggedArtifact-get_metadata}{}}}
\subsection{Method \code{get_metadata()}}{
Get the metadata of the artifact.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoggedArtifact$get_metadata()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoggedArtifact-get_version_tags"></a>}}
\if{latex}{\out{\hypertarget{method-LoggedArtifact-get_version_tags}{}}}
\subsection{Method \code{get_version_tags()}}{
Get the list of tags of the artifact version.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoggedArtifact$get_version_tags()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoggedArtifact-assets"></a>}}
\if{latex}{\out{\hypertarget{method-LoggedArtifact-assets}{}}}
\subsection{Method \code{assets()}}{
Get the list of all \code{\link{LoggedArtifactAsset}} that have been logged with
this \code{LoggedArtifact}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoggedArtifact$assets()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-LoggedArtifact-remote_assets"></a>}}
\if{latex}{\out{\hypertarget{method-LoggedArtifact-remote_assets}{}}}
\subsection{Method \code{remote_assets()}}{
Get the list of remote \code{\link{LoggedArtifactAsset}} that have been logged
with this \code{LoggedArtifact}.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{LoggedArtifact$remote_assets()}\if{html}{\out{</div>}}
}

}
}
