% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_general.R
\name{projects}
\alias{projects}
\title{Get a workspace's projects}
\usage{
projects(workspace_name = NULL, api_key = NULL)
}
\arguments{
\item{workspace_name}{Workspace name.}

\item{api_key}{Comet API key.}
}
\description{
Get a workspace's projects
}
