% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/read_general.R
\name{workspaces}
\alias{workspaces}
\title{Get a user's workspaces}
\usage{
workspaces(api_key = NULL)
}
\arguments{
\item{api_key}{Comet API key (can also be specified using the \code{COMET_API_KEY}
parameter as an environment variable or in a comet config file).}
}
\description{
Get a user's workspaces
}
