% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{create_experiment}
\alias{create_experiment}
\title{Create a new experiment}
\usage{
create_experiment(
  experiment_name = NULL,
  project_name = NULL,
  workspace_name = NULL,
  api_key = NULL,
  log_stderr = FALSE
)
}
\arguments{
\item{experiment_name}{Experiment name.}

\item{project_name}{Project name (can also be specified using the \code{COMET_PROJECT_NAME}
parameter as an environment variable or in a comet config file).}

\item{workspace_name}{Workspace name (can also be specified using the \code{COMET_WORKSPACE}
parameter as an environment variable or in a comet config file).}

\item{api_key}{Comet API key (can also be specified using the \code{COMET_API_KEY}
parameter as an environment variable or in a comet config file).}

\item{log_stderr}{If \code{TRUE}, all output from 'stderr' (which includes errors,
warnings, and messages) will be redirected to the Comet servers to display as message
logs for the experiment. If \code{FALSE}, 'stderr' will be directed to its default handler
but its output will not be logged with the experiment.}
}
\value{
An \code{\link{Experiment}} object.
}
\description{
Create a new experiment on Comet's servers. The return value is an \code{\link{Experiment}}
object that can be used to modify or get information about the experiment. Only one
experiment can be active at a time, so make sure to stop an experiment before creating
a new one (by calling the \code{stop()} method on the \code{\link{Experiment}} object).
}
