% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{experiment}
\alias{experiment}
\title{Create a new experiment}
\usage{
experiment(
  project_name = NULL,
  workspace_name = NULL,
  api_key = NULL,
  log_errors = FALSE
)
}
\arguments{
\item{project_name}{Project name (can also be specified using the \code{COMET_PROJECT_NAME}
parameter as an environment variable or in a comet config file).}

\item{workspace_name}{Workspace name (can also be specified using the \code{COMET_WORKSPACE}
parameter as an environment variable or in a comet config file).}

\item{api_key}{Comet API key (can also be specified using the \code{COMET_API_KEY}
parameter as an environment variable or in a comet config file).}

\item{log_errors}{Whether or not to log errors.}
}
\description{
Create a new experiment
}
