% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{Experiment}
\alias{Experiment}
\title{A Comet Experiment object}
\description{
A comet experiment object can be used to modify or get information about an active
experiment. All methods documented here are the different ways to interact with an
experiment. Use \code{\link[=create_experiment]{create_experiment()}} to create a Comet experiment object.
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Experiment$new()}}
\item \href{#method-get_experiment_key}{\code{Experiment$get_experiment_key()}}
\item \href{#method-get_experiment_url}{\code{Experiment$get_experiment_url()}}
\item \href{#method-get_html}{\code{Experiment$get_html()}}
\item \href{#method-log_html}{\code{Experiment$log_html()}}
\item \href{#method-get_asset_list}{\code{Experiment$get_asset_list()}}
\item \href{#method-get_asset}{\code{Experiment$get_asset()}}
\item \href{#method-upload_asset}{\code{Experiment$upload_asset()}}
\item \href{#method-get_graph}{\code{Experiment$get_graph()}}
\item \href{#method-log_graph}{\code{Experiment$log_graph()}}
\item \href{#method-get_tags}{\code{Experiment$get_tags()}}
\item \href{#method-add_tags}{\code{Experiment$add_tags()}}
\item \href{#method-get_other}{\code{Experiment$get_other()}}
\item \href{#method-log_other}{\code{Experiment$log_other()}}
\item \href{#method-create_symlink}{\code{Experiment$create_symlink()}}
\item \href{#method-get_output}{\code{Experiment$get_output()}}
\item \href{#method-get_code}{\code{Experiment$get_code()}}
\item \href{#method-get_system_details}{\code{Experiment$get_system_details()}}
\item \href{#method-archive}{\code{Experiment$archive()}}
\item \href{#method-restore}{\code{Experiment$restore()}}
\item \href{#method-delete}{\code{Experiment$delete()}}
\item \href{#method-stop}{\code{Experiment$stop()}}
\item \href{#method-print}{\code{Experiment$print()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Do not call this function directly. Use \code{create_experiment()} instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$new(
  experiment_key,
  experiment_url = NULL,
  api_key = NULL,
  keep_active = TRUE,
  log_output = TRUE,
  log_error = FALSE
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{experiment_key}}{N/A}

\item{\code{experiment_url}}{N/A}

\item{\code{api_key}}{N/A}

\item{\code{keep_active}}{N/A}

\item{\code{log_output}}{N/A}

\item{\code{log_error}}{N/A}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_experiment_key"></a>}}
\if{latex}{\out{\hypertarget{method-get_experiment_key}{}}}
\subsection{Method \code{get_experiment_key()}}{
Get the experiment key of an experiment.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$get_experiment_key()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_experiment_url"></a>}}
\if{latex}{\out{\hypertarget{method-get_experiment_url}{}}}
\subsection{Method \code{get_experiment_url()}}{
Get the URL to view an experiment in the browser.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$get_experiment_url()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_html"></a>}}
\if{latex}{\out{\hypertarget{method-get_html}{}}}
\subsection{Method \code{get_html()}}{
Get an experiment's HTML.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$get_html()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-log_html"></a>}}
\if{latex}{\out{\hypertarget{method-log_html}{}}}
\subsection{Method \code{log_html()}}{
Set (or append onto) an experiment's HTML.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$log_html(html, override = FALSE)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{html}}{(Required) An HTML string to add to the experiment.}

\item{\code{override}}{If \code{TRUE}, override the previous HTML. If \code{FALSE}, append to it.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_asset_list"></a>}}
\if{latex}{\out{\hypertarget{method-get_asset_list}{}}}
\subsection{Method \code{get_asset_list()}}{
Get an experiment's asset list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$get_asset_list(type = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{The type of assets to retrieve (by default, all assets are returned).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_asset"></a>}}
\if{latex}{\out{\hypertarget{method-get_asset}{}}}
\subsection{Method \code{get_asset()}}{
Get an asset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$get_asset(assetId)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{assetId}}{(Required) The asset ID to retrieve.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-upload_asset"></a>}}
\if{latex}{\out{\hypertarget{method-upload_asset}{}}}
\subsection{Method \code{upload_asset()}}{
Upload a file to the experiment.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$upload_asset(
  file,
  step = NULL,
  overwrite = NULL,
  context = NULL,
  type = NULL,
  name = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{(Required) Path to the file to upload.}

\item{\code{step}}{Step number.}

\item{\code{overwrite}}{If \code{TRUE}, overwrite any uploaded file with the same name.}

\item{\code{context}}{The context.}

\item{\code{type}}{The type of asset.}

\item{\code{name}}{Name of the file on comet. By default the name of the file will
match the file that you upload, but you can use this parameter to use a
different name.}

\item{\code{metadata}}{Metadata to upload along with the file.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_graph"></a>}}
\if{latex}{\out{\hypertarget{method-get_graph}{}}}
\subsection{Method \code{get_graph()}}{
Get an experiment's model graph.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$get_graph()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-log_graph"></a>}}
\if{latex}{\out{\hypertarget{method-log_graph}{}}}
\subsection{Method \code{log_graph()}}{
Log an experiment's associated model graph.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$log_graph(graph)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{graph}}{JSON representation of a graph.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_tags"></a>}}
\if{latex}{\out{\hypertarget{method-get_tags}{}}}
\subsection{Method \code{get_tags()}}{
Get an experiment's tags.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$get_tags()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-add_tags"></a>}}
\if{latex}{\out{\hypertarget{method-add_tags}{}}}
\subsection{Method \code{add_tags()}}{
Add a list of tags to an experiment.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$add_tags(tags)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{tags}}{List of tags.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_other"></a>}}
\if{latex}{\out{\hypertarget{method-get_other}{}}}
\subsection{Method \code{get_other()}}{
Get an experiment's others (logged with \code{log_other()}) summary.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$get_other()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-log_other"></a>}}
\if{latex}{\out{\hypertarget{method-log_other}{}}}
\subsection{Method \code{log_other()}}{
Log a key/value `other`` data (not a metric or parameter). Note
that you can only retrieve others summary data (e.g., this is
not recorded as a full time series).
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$log_other(key, value)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{key}}{The key.}

\item{\code{value}}{The value.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-create_symlink"></a>}}
\if{latex}{\out{\hypertarget{method-create_symlink}{}}}
\subsection{Method \code{create_symlink()}}{
Add a symlink to an experiment in another project.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$create_symlink(project_name)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{project_name}}{(Required) Project that the experiment to should linked to.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_output"></a>}}
\if{latex}{\out{\hypertarget{method-get_output}{}}}
\subsection{Method \code{get_output()}}{
Get an experiment's standard output and error.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$get_output()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_code"></a>}}
\if{latex}{\out{\hypertarget{method-get_code}{}}}
\subsection{Method \code{get_code()}}{
Get an experiment's source code.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$get_code()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_system_details"></a>}}
\if{latex}{\out{\hypertarget{method-get_system_details}{}}}
\subsection{Method \code{get_system_details()}}{
Get an experiment's system details.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$get_system_details()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-archive"></a>}}
\if{latex}{\out{\hypertarget{method-archive}{}}}
\subsection{Method \code{archive()}}{
Archive an experiment.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$archive()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-restore"></a>}}
\if{latex}{\out{\hypertarget{method-restore}{}}}
\subsection{Method \code{restore()}}{
Restore an archived experiment.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$restore()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-delete"></a>}}
\if{latex}{\out{\hypertarget{method-delete}{}}}
\subsection{Method \code{delete()}}{
Delete an experiment.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$delete()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stop"></a>}}
\if{latex}{\out{\hypertarget{method-stop}{}}}
\subsection{Method \code{stop()}}{
Stop an experiment. Always call this method before creating a new experiment.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$stop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the experiment.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$print()}\if{html}{\out{</div>}}
}

}
}
