% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{experiment}
\alias{experiment}
\title{Create a new experiment}
\usage{
experiment(
  project_name = NULL,
  workspace_name = NULL,
  api_key = NULL,
  log_errors = FALSE
)
}
\arguments{
\item{project_name}{Project name.}

\item{workspace_name}{Workspace name.}

\item{api_key}{Comet API key.}

\item{log_errors}{Whether or not to log errors.}
}
\description{
Create a new experiment
}
