% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{Experiment}
\alias{Experiment}
\title{vv}
\description{
pp
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Experiment$new()}}
\item \href{#method-stop}{\code{Experiment$stop()}}
\item \href{#method-get_experiment_key}{\code{Experiment$get_experiment_key()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
sdfds
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$new(experiment_key)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{experiment_key}}{dsf}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stop"></a>}}
\if{latex}{\out{\hypertarget{method-stop}{}}}
\subsection{Method \code{stop()}}{
Stop an experiment
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$stop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_experiment_key"></a>}}
\if{latex}{\out{\hypertarget{method-get_experiment_key}{}}}
\subsection{Method \code{get_experiment_key()}}{
haha
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$get_experiment_key()}\if{html}{\out{</div>}}
}

\subsection{Returns}{
the key
}
}
}
