% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/experiment.R
\name{Experiment}
\alias{Experiment}
\title{vv}
\description{
pp
}
\section{Methods}{
\subsection{Public methods}{
\itemize{
\item \href{#method-new}{\code{Experiment$new()}}
\item \href{#method-print}{\code{Experiment$print()}}
\item \href{#method-stop}{\code{Experiment$stop()}}
\item \href{#method-get_experiment_key}{\code{Experiment$get_experiment_key()}}
\item \href{#method-get_html}{\code{Experiment$get_html()}}
\item \href{#method-set_html}{\code{Experiment$set_html()}}
\item \href{#method-get_asset_list}{\code{Experiment$get_asset_list()}}
\item \href{#method-get_asset}{\code{Experiment$get_asset()}}
\item \href{#method-upload_asset}{\code{Experiment$upload_asset()}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-new"></a>}}
\if{latex}{\out{\hypertarget{method-new}{}}}
\subsection{Method \code{new()}}{
Do not call this function directly. Use \code{create_experiment()} instead.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$new(experiment_key, log_stderr = FALSE, api_key = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{experiment_key}}{.}

\item{\code{log_stderr}}{.}

\item{\code{api_key}}{.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-print"></a>}}
\if{latex}{\out{\hypertarget{method-print}{}}}
\subsection{Method \code{print()}}{
Print the experiment.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$print()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-stop"></a>}}
\if{latex}{\out{\hypertarget{method-stop}{}}}
\subsection{Method \code{stop()}}{
Stop an experiment.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$stop()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_experiment_key"></a>}}
\if{latex}{\out{\hypertarget{method-get_experiment_key}{}}}
\subsection{Method \code{get_experiment_key()}}{
Get the experiment key of an experiment.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$get_experiment_key()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_html"></a>}}
\if{latex}{\out{\hypertarget{method-get_html}{}}}
\subsection{Method \code{get_html()}}{
Get an experiment's HTML.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$get_html()}\if{html}{\out{</div>}}
}

}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-set_html"></a>}}
\if{latex}{\out{\hypertarget{method-set_html}{}}}
\subsection{Method \code{set_html()}}{
Set (or append onto) an experiment's HTML.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$set_html(html, override = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{html}}{(Required) An HTML string to add to the experiment.}

\item{\code{override}}{If \code{TRUE}, override the previous HTML. If \code{FALSE}, append to it.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_asset_list"></a>}}
\if{latex}{\out{\hypertarget{method-get_asset_list}{}}}
\subsection{Method \code{get_asset_list()}}{
Get an experiment's asset list.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$get_asset_list(type = NULL)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{type}}{The type of assets to retrieve (by default, all assets are returned).}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-get_asset"></a>}}
\if{latex}{\out{\hypertarget{method-get_asset}{}}}
\subsection{Method \code{get_asset()}}{
Get an asset.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$get_asset(assetId)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{assetId}}{(Required) The asset ID to retrieve.}
}
\if{html}{\out{</div>}}
}
}
\if{html}{\out{<hr>}}
\if{html}{\out{<a id="method-upload_asset"></a>}}
\if{latex}{\out{\hypertarget{method-upload_asset}{}}}
\subsection{Method \code{upload_asset()}}{
Upload a file to the experiment.
\subsection{Usage}{
\if{html}{\out{<div class="r">}}\preformatted{Experiment$upload_asset(
  file,
  step = NULL,
  overwrite = NULL,
  context = NULL,
  type = NULL,
  name = NULL,
  metadata = NULL
)}\if{html}{\out{</div>}}
}

\subsection{Arguments}{
\if{html}{\out{<div class="arguments">}}
\describe{
\item{\code{file}}{(Required) Path to the file to upload.}

\item{\code{step}}{Step number.}

\item{\code{overwrite}}{If \code{TRUE}, overwrite any uploaded file with the same name.}

\item{\code{context}}{The context.}

\item{\code{type}}{The type of asset.}

\item{\code{name}}{Name of the file on comet. By default the name of the file will
match the file that you upload, but you can use this parameter to use a
different name.}

\item{\code{metadata}}{Metadata to upload along with the file.}
}
\if{html}{\out{</div>}}
}
}
}
