% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/general-methods.R
\name{get_projects}
\alias{get_projects}
\title{Get a workspace's projects}
\usage{
get_projects(workspace_name = NULL, api_key = NULL)
}
\arguments{
\item{workspace_name}{Workspace name (can also be specified using the \code{COMET_WORKSPACE}
parameter as an environment variable or in a comet config file).}

\item{api_key}{Comet API key (can also be specified using the \code{COMET_API_KEY}
parameter as an environment variable or in a comet config file).}
}
\description{
Get a workspace's projects
}
\examples{
\dontrun{
library(cometr)
# Assuming you have COMET_API_KEY, COMET_WORKSPACE variables defined
get_projects()
}

}
